refactor: MVC 패턴으로 리팩토링
