+++ src/test/java/racingcar/view/RacingGameInputTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.view;
 
 import org.junit.jupiter.api.Test;
